gateway-api: extract secret-sync related helper functions
This commit extracts the secret-sync related Gateway API
functions into a dedicated file.

This way the secret-sync can be moved into its own package.
