+++ operator/pkg/gateway-api/secret.go
@@ -16,10 +16,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
-
-	"github.com/cilium/cilium/operator/pkg/gateway-api/helpers"
-	"github.com/cilium/cilium/pkg/logging/logfields"
 )
 
 const (
@@ -66,44 +62,6 @@ func (r *secretSyncer) notInSecretsNamespace() builder.Predicates {
 	return builder.WithPredicates(predicate.NewPredicateFuncs(func(object client.Object) bool {
 		return object.GetNamespace() != r.secretsNamespace
 	}))
-}
-
-func enqueueTLSSecrets(c client.Client) handler.EventHandler {
-	return handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, obj client.Object) []reconcile.Request {
-		scopedLog := log.WithContext(ctx).WithFields(logrus.Fields{
-			logfields.Controller: "secrets",
-			logfields.Resource:   obj.GetName(),
-		})
-
-		gw, ok := obj.(*gatewayv1.Gateway)
-		if !ok {
-			return nil
-		}
-
-		// Check whether Gateway is managed by Cilium
-		if !hasMatchingController(ctx, c, controllerName)(gw) {
-			return nil
-		}
-
-		var reqs []reconcile.Request
-		for _, l := range gw.Spec.Listeners {
-			if l.TLS == nil {
-				continue
-			}
-			for _, cert := range l.TLS.CertificateRefs {
-				if !helpers.IsSecret(cert) {
-					continue
-				}
-				s := types.NamespacedName{
-					Namespace: helpers.NamespaceDerefOr(cert.Namespace, gw.Namespace),
-					Name:      string(cert.Name),
-				}
-				reqs = append(reqs, reconcile.Request{NamespacedName: s})
-				scopedLog.WithField("secret", s).Debug("Enqueued secret for gateway")
-			}
-		}
-		return reqs
-	})
 }
 
 func enqueueOwningSecretFromLabels() handler.EventHandler {
