+++ operator/pkg/gateway-api/secretsync.go
@@ -1,0 +1,68 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+package gateway_api
+
+import (
+	"context"
+
+	"github.com/sirupsen/logrus"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
+
+	"github.com/cilium/cilium/operator/pkg/gateway-api/helpers"
+	"github.com/cilium/cilium/pkg/logging/logfields"
+)
+
+func enqueueTLSSecrets(c client.Client) handler.EventHandler {
+	return handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, obj client.Object) []reconcile.Request {
+		scopedLog := log.WithContext(ctx).WithFields(logrus.Fields{
+			logfields.Controller: "secrets",
+			logfields.Resource:   obj.GetName(),
+		})
+
+		gw, ok := obj.(*gatewayv1.Gateway)
+		if !ok {
+			return nil
+		}
+
+		// Check whether Gateway is managed by Cilium
+		if !hasMatchingController(ctx, c, controllerName)(gw) {
+			return nil
+		}
+
+		var reqs []reconcile.Request
+		for _, l := range gw.Spec.Listeners {
+			if l.TLS == nil {
+				continue
+			}
+			for _, cert := range l.TLS.CertificateRefs {
+				if !helpers.IsSecret(cert) {
+					continue
+				}
+				s := types.NamespacedName{
+					Namespace: helpers.NamespaceDerefOr(cert.Namespace, gw.Namespace),
+					Name:      string(cert.Name),
+				}
+				reqs = append(reqs, reconcile.Request{NamespacedName: s})
+				scopedLog.WithField("secret", s).Debug("Enqueued secret for gateway")
+			}
+		}
+		return reqs
+	})
+}
+
+func isUsedByCiliumGateway(ctx context.Context, c client.Client, obj *corev1.Secret) bool {
+	gateways := getGatewaysForSecret(ctx, c, obj)
+	for _, gw := range gateways {
+		if hasMatchingController(ctx, c, controllerName)(gw) {
+			return true
+		}
+	}
+
+	return false
+}
